requirements.txt
gql==3.5.0
aiohttp==3.9.1
psycopg2-binary==2.9.9
aiosqlite==0.19.0
python-dotenv==1.0.0